fix: tests and rename api.devices.spec to api.boxes.$deviceId.spec
